Remove all use of nframes_t.
git-svn-id: svn://localhost/ardour2/branches/3.0@8166 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/evoral/src/Event.cpp
@@ -93,7 +93,7 @@ Event<Timestamp>::~Event() {
 #endif // EVORAL_EVENT_ALLOC
 
 template class Event<Evoral::MusicalTime>;
-template class Event<uint32_t>;
+template class Event<int64_t>;
 
 } // namespace Evoral
